complete routing of work excluding Turkson's pages
