+++ m-fix/package.json
@@ -10,17 +10,26 @@
   },
   "dependencies": {
     "@expo/metro-runtime": "~5.0.4",
+    "@react-native-community/masked-view": "^0.1.11",
+    "@react-native-picker/picker": "^2.11.0",
+    "@react-navigation/bottom-tabs": "^7.3.14",
+    "@react-navigation/native": "^7.1.10",
+    "@react-navigation/native-stack": "^7.3.14",
     "expo": "~53.0.9",
+    "expo-av": "~15.1.4",
     "expo-status-bar": "~2.2.3",
+    "expo-video": "~2.1.9",
+    "prop-types": "^15.8.1",
     "react": "19.0.0",
     "react-dom": "19.0.0",
     "react-native": "0.79.2",
-    "react-native-gesture-handler": "^2.25.0",
-    "react-native-web": "^0.20.0"
+    "react-native-gesture-handler": "~2.24.0",
+    "react-native-picker-select": "^9.3.1",
+    "react-native-reanimated": "~3.17.4",
+    "react-native-safe-area-context": "5.4.0",
+    "react-native-screens": "~4.10.0",
     "react-native-video": "^6.14.1",
-    "expo-av": "~15.1.4",
-    "expo-video": "~2.1.9",
-    "undefined": "react-native-vector-icons/Feather"
+    "react-native-web": "^0.20.0"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
