+++ m-fix/screens/auth/Register.tsx
@@ -10,6 +10,7 @@ import {
 } from 'react-native';
 import RNPickerSelect from 'react-native-picker-select';
 
+  
 export default function SignUpScreen() {
   const [code, setCode] = useState('');
   const [password, setPassword] = useState('');
